Construct SimpleDarwinRecord class dynamically. (#84)
* Construct SimpleDarwinRecord class dynamically.

Use a list of DWC terms to create the fields in the SimpleDarwinRecord
class. This works because all the fields are StringFields, and we use
consistent rules to name the class fields from the DWC terms. In the
future, to add more of the DWC fields, we just need to add the name of
the term to the list.

* fix to work on python 2